fix: add trusted-contact verification routing to IPC handler
