+++ assistant/src/runtime/routes/integration-routes.ts
@@ -21,14 +21,12 @@
  * GET    /v1/channel-verification-sessions/status   — check guardian binding status
  */
 
-import { createHash, randomBytes } from "node:crypto";
-
 import type { ChannelId } from "../../channels/types.js";
-import { getChannelById, getContact } from "../../contacts/contact-store.js";
 import {
   createInboundChallenge,
   getVerificationStatus,
   revokeVerificationForChannel,
+  verifyTrustedContact,
 } from "../../daemon/handlers/config-channels.js";
 import {
   clearSlackChannelConfig,
@@ -42,32 +40,17 @@ import {
   setTelegramConfig,
   setupTelegram,
 } from "../../daemon/handlers/config-telegram.js";
-import { getCredentialMetadata } from "../../tools/credentials/metadata-store.js";
 import { normalizePhoneNumber } from "../../util/phone.js";
-import {
-  countRecentSendsToDestination,
-  createOutboundSession,
-  revokePendingSessions,
-  updateSessionDelivery,
-} from "../channel-verification-service.js";
+import { revokePendingSessions } from "../channel-verification-service.js";
 import { httpError } from "../http-errors.js";
 import type { RouteDefinition } from "../http-router.js";
 import {
   cancelOutbound,
-  deliverVerificationSlack,
-  deliverVerificationTelegram,
-  DESTINATION_RATE_WINDOW_MS,
-  MAX_SENDS_PER_DESTINATION_WINDOW,
   normalizeTelegramDestination,
   resendOutbound,
   startOutbound,
 } from "../verification-outbound-actions.js";
 import { guardianVerificationLimiter } from "../verification-rate-limiter.js";
-import {
-  composeVerificationSlack,
-  composeVerificationTelegram,
-  GUARDIAN_VERIFY_TEMPLATE_KEYS,
-} from "../verification-templates.js";
 
 /**
  * GET /v1/integrations/telegram/config
@@ -166,42 +149,6 @@ export async function handleClearSlackChannelConfig(): Promise<Response> {
 // Channel verification (unified session API)
 // ---------------------------------------------------------------------------
 
-/** Session TTL in seconds (matches challenge TTL of 10 minutes). */
-const SESSION_TTL_SECONDS = 600;
-
-/**
- * Map a contact channel type to the verification ChannelId used by the
- * verification service. Returns null for unsupported channel types.
- */
-function toVerificationChannel(channelType: string): ChannelId | null {
-  switch (channelType) {
-    case "phone":
-      return "voice";
-    case "telegram":
-      return "telegram";
-    case "slack":
-      return "slack";
-    default:
-      return null;
-  }
-}
-
-/**
- * Get the Telegram bot username from credential metadata.
- * Falls back to process.env.TELEGRAM_BOT_USERNAME.
- */
-function getTelegramBotUsername(): string | undefined {
-  const meta = getCredentialMetadata("telegram", "bot_token");
-  if (
-    meta?.accountInfo &&
-    typeof meta.accountInfo === "string" &&
-    meta.accountInfo.trim().length > 0
-  ) {
-    return meta.accountInfo.trim();
-  }
-  return process.env.TELEGRAM_BOT_USERNAME || undefined;
-}
-
 /**
  * POST /v1/channel-verification-sessions
  *
@@ -228,10 +175,21 @@ export async function handleCreateVerificationSession(
 
   const purpose = body.purpose ?? "guardian";
 
-  // Trusted contact verification path — look up the contact channel and derive
-  // channel/destination from it, then create a session with verificationPurpose: "trusted_contact".
+  // Trusted contact verification path — delegates to the shared transport-agnostic
+  // function and wraps the result in an HTTP response.
   if (purpose === "trusted_contact" && body.contactChannelId) {
-    return handleTrustedContactVerification(body.contactChannelId, assistantId);
+    const result = await verifyTrustedContact(
+      body.contactChannelId,
+      assistantId,
+    );
+    const status = result.success
+      ? 200
+      : result.error === "rate_limited"
+        ? 429
+        : result.error === "already_verified"
+          ? 409
+          : 400;
+    return Response.json(result, { status });
   }
 
   if (body.destination) {
@@ -286,205 +244,6 @@ export async function handleCreateVerificationSession(
   );
   const status = result.success ? 200 : 400;
   return Response.json(result, { status });
-}
-
-/**
- * Trusted contact verification — extracted from the former
- * handleVerifyContactChannel in contact-routes.ts. Looks up the contact
- * channel, derives channelId and destination, checks rate limits, and
- * creates the session with verificationPurpose: "trusted_contact".
- */
-async function handleTrustedContactVerification(
-  contactChannelId: string,
-  assistantId: string,
-): Promise<Response> {
-  const channel = getChannelById(contactChannelId);
-  if (!channel) {
-    return httpError(
-      "NOT_FOUND",
-      `Channel "${contactChannelId}" not found`,
-      404,
-    );
-  }
-
-  const contact = getContact(channel.contactId);
-  if (!contact) {
-    return httpError(
-      "NOT_FOUND",
-      `Contact "${channel.contactId}" not found`,
-      404,
-    );
-  }
-
-  // Already verified — no need to re-verify
-  if (channel.status === "active" && channel.verifiedAt != null) {
-    return httpError("CONFLICT", "Channel is already verified", 409);
-  }
-
-  const verificationChannel = toVerificationChannel(channel.type);
-  if (!verificationChannel) {
-    return httpError(
-      "BAD_REQUEST",
-      `Verification is not supported for channel type "${channel.type}"`,
-      400,
-    );
-  }
-
-  const destination = channel.address;
-  if (!destination) {
-    return httpError(
-      "BAD_REQUEST",
-      "Channel has no address to send verification to",
-      400,
-    );
-  }
-
-  // Normalize Telegram destinations so rate-limit lookups are consistent
-  const effectiveDestination =
-    verificationChannel === "telegram"
-      ? normalizeTelegramDestination(destination)
-      : destination;
-
-  // Rate limit check
-  const recentSendCount = countRecentSendsToDestination(
-    verificationChannel,
-    effectiveDestination,
-    DESTINATION_RATE_WINDOW_MS,
-  );
-  if (recentSendCount >= MAX_SENDS_PER_DESTINATION_WINDOW) {
-    return httpError(
-      "RATE_LIMITED",
-      "Too many verification attempts to this destination. Please try again later.",
-      429,
-    );
-  }
-
-  // --- Telegram verification ---
-  if (verificationChannel === "telegram") {
-    // Telegram with known chat ID: identity is already bound
-    if (channel.externalChatId) {
-      const sessionResult = createOutboundSession({
-        channel: verificationChannel,
-        expectedChatId: channel.externalChatId,
-        expectedExternalUserId: channel.externalUserId ?? undefined,
-        identityBindingStatus: "bound",
-        destinationAddress: effectiveDestination,
-        verificationPurpose: "trusted_contact",
-      });
-
-      const telegramBody = composeVerificationTelegram(
-        GUARDIAN_VERIFY_TEMPLATE_KEYS.TELEGRAM_CHALLENGE_REQUEST,
-        {
-          code: sessionResult.secret,
-          expiresInMinutes: Math.floor(SESSION_TTL_SECONDS / 60),
-        },
-      );
-
-      const now = Date.now();
-      const sendCount = 1;
-      updateSessionDelivery(sessionResult.sessionId, now, sendCount, null);
-      deliverVerificationTelegram(
-        channel.externalChatId,
-        telegramBody,
-        assistantId,
-      );
-
-      return Response.json({
-        ok: true,
-        verificationSessionId: sessionResult.sessionId,
-        expiresAt: sessionResult.expiresAt,
-        sendCount,
-      });
-    }
-
-    // Telegram handle only (no chat ID): bootstrap flow
-    const { ensureTelegramBotUsernameResolved } =
-      await import("../channel-invite-transports/telegram.js");
-    await ensureTelegramBotUsernameResolved();
-    const botUsername = getTelegramBotUsername();
-    if (!botUsername) {
-      return httpError(
-        "BAD_REQUEST",
-        "Telegram bot username is not configured. Set up the Telegram integration first.",
-        400,
-      );
-    }
-
-    const bootstrapToken = randomBytes(16).toString("hex");
-    const bootstrapTokenHash = createHash("sha256")
-      .update(bootstrapToken)
-      .digest("hex");
-
-    const sessionResult = createOutboundSession({
-      channel: verificationChannel,
-      identityBindingStatus: "pending_bootstrap",
-      destinationAddress: effectiveDestination,
-      bootstrapTokenHash,
-      verificationPurpose: "trusted_contact",
-    });
-
-    const telegramBootstrapUrl = `https://t.me/${botUsername}?start=gv_${bootstrapToken}`;
-
-    return Response.json({
-      ok: true,
-      verificationSessionId: sessionResult.sessionId,
-      expiresAt: sessionResult.expiresAt,
-      sendCount: 0,
-      telegramBootstrapUrl,
-    });
-  }
-
-  // --- Slack verification ---
-  if (verificationChannel === "slack") {
-    const slackUserId = channel.externalUserId ?? destination;
-
-    // Only claim identity is bound when we have at least one platform identifier
-    const hasIdentityBinding = Boolean(
-      channel.externalUserId || channel.externalChatId,
-    );
-    if (!hasIdentityBinding) {
-      return httpError(
-        "BAD_REQUEST",
-        "Slack verification requires an externalUserId or externalChatId for identity binding",
-        400,
-      );
-    }
-
-    const sessionResult = createOutboundSession({
-      channel: verificationChannel,
-      expectedExternalUserId: channel.externalUserId ?? undefined,
-      expectedChatId: channel.externalChatId ?? undefined,
-      identityBindingStatus: "bound",
-      destinationAddress: slackUserId,
-      verificationPurpose: "trusted_contact",
-    });
-
-    const slackBody = composeVerificationSlack(
-      GUARDIAN_VERIFY_TEMPLATE_KEYS.SLACK_CHALLENGE_REQUEST,
-      {
-        code: sessionResult.secret,
-        expiresInMinutes: Math.floor(SESSION_TTL_SECONDS / 60),
-      },
-    );
-
-    const now = Date.now();
-    const sendCount = 1;
-    updateSessionDelivery(sessionResult.sessionId, now, sendCount, null);
-    deliverVerificationSlack(slackUserId, slackBody, assistantId);
-
-    return Response.json({
-      ok: true,
-      verificationSessionId: sessionResult.sessionId,
-      expiresAt: sessionResult.expiresAt,
-      sendCount,
-    });
-  }
-
-  return httpError(
-    "BAD_REQUEST",
-    `Verification is not supported for channel type "${channel.type}"`,
-    400,
-  );
 }
 
 /**
